12-06-2022

+++ database/migrations/2022_05_29_054340_create_sub_descriptions_table.php
@@ -25,7 +25,7 @@ class CreateSubDescriptionsTable extends Migration
             $table->date('financial_opening_date')->nullable();
             $table->tinyInteger('technical_weightage')->unsigned()->nullable();
             $table->tinyInteger('financial_weightage')->unsigned()->nullable();
-            $table->tinyInteger('passing_marks')->unsigned()->nullable();
+            $table->decimal('passing_marks',4,0)->unsigned()->nullable();
             $table->timestamps();
             $table->foreign('submission_id')->references('id')->on('submissions')->onDelete('cascade');
             $table->foreign('sub_status_id')->references('id')->on('sub_statuses');
